# Conflicts:
#	features/content/play_broadcaster/src/main/java/com/tokopedia/play/broadcaster/view/activity/PlayBroadcastActivity.kt

+++ features/merchant/official_store/src/main/java/com/tokopedia/officialstore/analytics/OfficialStoreTracking.kt
@@ -696,7 +696,7 @@ class OfficialStoreTracking(context: Context) {
         return "/official-store$stringIsLogin - rekomendasi untuk anda - $recommendationType$stringTopAds"
     }
 
-    fun eventClickWishlist(categoryName: String, isAddWishlist: Boolean, isLogin: Boolean, productId: Int, isTopAds: Boolean) {
+    fun eventClickWishlist(categoryName: String, isAddWishlist: Boolean, isLogin: Boolean, productId: Long, isTopAds: Boolean) {
         val action = if (isAddWishlist) "add" else "remove"
         val statusTopads = if (isTopAds) "topads" else  "general"
         var eventAction = "$action wishlist - product recommendation - ${if (isLogin) "login" else "non login"}"
